Add audio feedback to Timer and minor UI tweaks
Added a beep sound to the Timer component, which is played during the final countdown. Also updated the Timer to change color and size as it approaches zero. Fixed the order of messages in the Chat component to display messages in chronological order. Corrected a text error and removed an unnecessary prop in the PlayerPage.